+++ src/Oro/Bundle/CMSBundle/Resources/public/js/app/grapesjs/controls/picture-settings/picture-settings-item-view.js
@@ -95,7 +95,7 @@ const PictureSettingsItemView = BaseView.extend({
                 title: __('oro.cms.wysiwyg.digital_asset.image.title'),
                 routeName: 'oro_digital_asset_widget_choose_image',
                 loadingElement: this.dialog.loadingElement,
-                onSelect: function(digitalAssetModel) {
+                onSelect(digitalAssetModel) {
                     model.updateImageUrl(digitalAssetModel.get('previewMetadata').url);
                 }
             }