+++ src/Oro/Bundle/CMSBundle/Tests/JS/types/picture-type-builderSpec.js
@@ -1,17 +1,20 @@
 import 'jasmine-jquery';
 import grapesJS from 'grapesjs';
-import PictureTypeBuilder from 'orocms/js/app/grapesjs/type-builders/picture-type-builder';
+import PictureTypeBuilder from 'orocms/js/app/grapesjs/types/picture-type';
 import ComponentRestriction from 'orocms/js/app/grapesjs/plugins/components/component-restriction';
 import html from 'text-loader!../fixtures/grapesjs-editor-view-fixture.html';
 
-describe('orocms/js/app/grapesjs/type-builders/image-type-builder', () => {
+describe('orocms/js/app/grapesjs/types/image', () => {
     let pictureTypeBuilder;
     let editor;
 
     beforeEach(done => {
         window.setFixtures(html);
         editor = grapesJS.init({
-            container: document.querySelector('.page-content-editor')
+            container: document.querySelector('.page-content-editor'),
+            deviceManager: {
+                devices: []
+            }
         });
 
         editor.ComponentRestriction = new ComponentRestriction(editor, {});