+++ src/Oro/Bundle/PricingBundle/Tests/Unit/Async/PriceListProcessorTest.php
@@ -190,7 +190,10 @@ class PriceListProcessorTest extends \PHPUnit\Framework\TestCase
         $dependentContext = $this->createMock(DependentJobContext::class);
         $dependentContext->expects($this->once())
             ->method('addDependentJob')
-            ->with(RunCombinedPriceListPostProcessingStepsTopic::getName(), ['relatedJobId' => 10]);
+            ->with(
+                RunCombinedPriceListPostProcessingStepsTopic::getName(),
+                ['relatedJobId' => 10, 'cpls' => [$cpl1->getId()]]
+            );
         $this->dependentJob->expects($this->once())
             ->method('createDependentJobContext')
             ->with($rootJob)
@@ -274,7 +277,10 @@ class PriceListProcessorTest extends \PHPUnit\Framework\TestCase
         $dependentContext = $this->createMock(DependentJobContext::class);
         $dependentContext->expects($this->once())
             ->method('addDependentJob')
-            ->with(RunCombinedPriceListPostProcessingStepsTopic::getName(), ['relatedJobId' => 10]);
+            ->with(
+                RunCombinedPriceListPostProcessingStepsTopic::getName(),
+                ['relatedJobId' => 10, 'cpls' => [10]]
+            );
         $this->dependentJob->expects($this->once())
             ->method('createDependentJobContext')
             ->with($rootJob)